Replace 'All Round Gothic' font with 'Rounded Elegance' for main logo (#160)
* Replace All Round Gothic with Rounded Elegance

This PR replaces 'All Round Gothic' with 'Rounded Elegance' for the site's logo font. The latter is free to use and I _think_ what Sam Simpson used on the original site. This is part one of a two-step process that will allow us to do away with https://github.com/audioxide/licensed-assets and have all Audioxide repositories public.

* Replace All Round Gothic with Rounded Elegance

This PR replaces 'All Round Gothic' with 'Rounded Elegance' for the site's logo font. The latter is free to use and I _think_ what Sam Simpson used on the original site. This is part one of a two-step process that will allow us to do away with https://github.com/audioxide/licensed-assets and have all Audioxide repositories public.

* Update package.json

* Move font files to webiste repo